About You in seperate components, cleaning up styles
